+++ external/debug_console.h
@@ -51,7 +51,7 @@ namespace con
 	typedef int64_t		i64;
 
 	typedef uint8_t		u8;
-	typedef uint16_t	u16;
+	typedef UINT	u16;
 	typedef uint32_t	u32;
 	typedef uint64_t	u64;
