+++ render/IndexBuffer.h
@@ -11,20 +11,20 @@ class IndexBuffer : public Bindable {
 public:
 	IndexBuffer();
 
-	IndexBuffer(std::vector<uint16_t> indexes); // maybe uint16_t too small for store indixes;
+	IndexBuffer(std::vector<UINT> indexes); // maybe UINT too small for store indixes;
 
 	~IndexBuffer();
 
-	void Create(std::vector<uint16_t> indexes);
+	void Create(std::vector<UINT> indexes);
 
 	void CreateDynamic(const UINT size);
 
-	void Update(std::vector<uint16_t> indexes);
+	void Update(std::vector<UINT> indexes);
 
 	virtual void Bind() override;
 	
-	uint16_t GetCount();
+	UINT GetCount();
 private:
-	uint16_t count;
+	UINT count;
 	ID3D11Buffer* buffer;
 };
