+++ render/UtilsRenderer.h
@@ -19,7 +19,7 @@
 
 struct UtilsDefinitions {
 	std::vector<UtilsVertex> vert;
-	std::vector<uint16_t> ind;
+	std::vector<UINT> ind;
 	MODEL cb;
 	int flags;
 };
